Updates required client version
This change updates the required client versions in our configuration
such that it will now support the updated 2014.2.1 code base.
This resolve issue where the service was uninstallable due to missing
client version.

This change also pinns the keystonemiddleware package as it was using
master and it should be set to using a stable tag.

The following clients were updated:
- cinderclient==1.1.1
- keystoneclient==1.0.0
- zaqarclient==0.1.0
- keystonemiddleware==1.3.1

To support this update we've also pinned global requirements to the
"stable/juno" branch.

Change-Id: I0c8546770f8e41fc42374785d3e7ba63f4d5adf9
Closes-Bug: 1416725

Conflicts:
	rpc_deployment/vars/repo_packages/keystonemiddleware.yml