Converts existing rsyslog role into rsyslog server
This commit changes the rsyslog role and what it was intended for from
within the environment map from a utility role to a server role. This
change makes it possible for rsyslog to recieve logs from services and
then store the agregated logs on a single host. The log directories will
be created dynamically in the `/var/log/rsyslog/` directory and will be
named after the host name that is shipping the logs.

Change-Id: Ida4fad84b0bf200cf902d3d35a9cfe7a7aebef9b