Adds new provider_networks module and reformats ml2_conf.ini
The ml2_conf.ini has been reformatted to work with the new data type
provided from the `provider_networks` module.

The new module parses and provides useful information from the
`provider_network` list as found within the base user config. The data
type has been documented within the defaults so that it can be used in
the role as stand alone, outside that of the OSAD project. The module
replaces functionality found in top of the ml2_conf.ini file that was
built using jinja2 syntax and flow control. The purpose is to allow
for more configuration possibilities while also being faster.

Closes-Bug: 1434942
Change-Id: If48267ba6f552eaf1535b8ecee6d26b2c6592b31