Backported tempest roles from Master
Change to backport the tempest roles from master into juno. This
change is being done because the deps within Juno no longer allow
tempest to be installed along side OpenStack without a lot of
tom foolery which is not worth the maintenance costs. The upstream
process for installing tempest work and will work with future
versions of the Juno development branch.

Change-Id: I18a690c614253d6b7e2e52e40bf1f14dd6dad87c