Moved user_group_vars to defaults
The change moves the user_group_vars.yml from openstack_deploy/ to
a standard default file. This will allow a deployer to override items
in the file while also allowing development to make changes to the file
that may impact its contents. When the file was moved out it was short
sighted to move it into 100% userspace and should have simply become the
main default file.

Change-Id: Icfc3efea0a43d55b5a23bca2f64d4d6d73542d10
Related-Bug: #1460516