Updated MariaDB to the new release version (10.0)
This change updates master to use the latest stable version of
MariaDB (move to 10.0 from our current 5.5). This change is to
open up new dev work in master and potentially other releases
which will allow us to evaluate the the new version of this
supporting service.

DocImpact
Implements: Blueprint MariaDB-upgrade-to-v10
Change-Id: I2d773e569667e07106862507b881e3f0aca85591