Break apart and document the upgrade process
This change breaks up the Juno>Kilo upgrade script such that its now
available to be called as separate pieces. This change allows the
deployer to call the various parts of the upgrade script as needed.
Special care has been taken to make the BASH parts of the script as
idempotent as possible

Documentation has been added to cover all of the scripts, what they
do, how they do it, and when to call them. While the `run-upgrade.sh`
script is the main entry point to doing an online upgrade, the script
itself is now simply an orchestrator which can be followed or
customized as needed.

The documentation should cover most, if not all, of the aspects of
an RPC Juno > OpenStack-Ansible Kilo upgrade. The deployer will
also have information on what we consider best practice and should
have the information required to make a decision on realistic
deliverables while migrating an old deployment to the new community
format which includes upgrading the installed OpenStack release as
well.

Change-Id: I82c347c1a5f911a95e58f502bb67bf10b2005b15