Updated the repo-build process
This commit changes the repo build play such that it no longer
requires "yaprt". The intention here is to make the build process
less of a blackbox, simpler, and faster. With recent improvements
in OpenStack and PIP we're now able to do a lot more without the
use of the "yaprt" tool. Repo building is now its own standalone
role.

This PR now reintroduces the repo-clone ability which is now made
smarter. The repo-build process creates a manifest file which is
used to provide the ability to sync exactly the contents of the
tagged release. This will make repeatable deployments faster for
downstream consumers.

The repo-store-source.yml play was removed because it hasn't been
used as the repo clone process is part of the main repo build
process.

DocImpact
UpgradeImpact
Change-Id: Icb0a4e3443491eb0242125ea59f8d5b50f89410b
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>

Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>