Added base-image building to the environment
This change adds the ability for the system to rebuild the
lxc base image from an upstream LXC rootfs as created by
"images.lxccontainers.org". The change creates the image
and builds a base index with the repo servers.

Once the image is created all subsequent container builds
are built using a cached image from the server to the hosts.
The cache will speed up container builds and consistency.

This change is a gateway to being able to do to an image
base deployment.

Change-Id: I7489381249f98fe536ef0844468b743f93683db1
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>