Added variable deprecation role and check
The change adds a role to check for variables that are in use which
have been deprecated. This change also brings within it an action
plugin with allows for the "warn" module to issues log entries for
items that are in use but have been deprecated. The new role can
be used as a dep to other roles and will react to the usage of
.*_fatal_deprecations being True.

The change was done in tree because it requires the use of an action
plugin to generate the messages. Action plugins are a server side
feature and can only carried in-tree.

The role was put into service with os_nova because its the only
role that has specifically called out items that are to be deprecated.

Change-Id: I7cd87050f140535a01bd1a4111eb82c91daab318
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>