Functional backport for the new repo-build process
The change backports the repo-build role into kilo such that
we're allowing the Kilo build process to benifit from the changes
we've made upstream. This will also allow for the kilo playbooks
to clone from a repo using a generated manifest file. This will
ensure that the build process is more repatable, creating an
environment with faster build times, and dropping our dependency
on the yaprt library.

This change brings with it all of the assosiated updates that have been
made to the various plugins and playbooks to ensure a better repo-env.

* In the change the repo-* plays have been backported from liberty
* the repo-build role has been backported
* the py_pkg lookup plugin changes has been backported

Change-Id: I5a3a1506bf3df8b5812d9131248d0156a8dd7dd6
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>