[WIP] Rewrote the Ansible SSH plugin
In Ansible 1.9.x there are a few areas where the SSH plugin
could / should be updated to better support the systems it
interacts with. This PR updated the SSH connection plugin
such that it is now more pythonic while also being smarter
in how the various methods write, update, and modifies data
in flight.

Change-Id: I60d4962d31024109f3aa610173fb3e068aaeadbb
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>